Added matrix multiplication from PMPP book

+++ Resources/Programming Massively Parallel Processors/matmul.cu
@@ -32,9 +32,9 @@ void MatrixMul(float* M, float* N,
     cudaMemcpy(N_d, N, size, cudaMemcpyHostToDevice);
     
     dim3 blockDim(32, 32);
-    dim3 gridDim(ceil(Width / (float)blockDim.x), ceil(Width / (float)blockDim.y));
+    dim3 gridDim(ceil(Width / blockDim.x), ceil(Width / blockDim.y));
 
-    printf("blockDim: (%d, %d, %d)\n", blockDim.x,blockDim.y, blockDim.z);
+    printf("blockDim: (%d, %d, %d)\n", blockDim.x, blockDim.y, blockDim.z);
     printf("gridDim: (%d, %d, %d)\n", gridDim.x, gridDim.y, gridDim.z);
 
     MatrixMulKernel<<<gridDim, blockDim>>>(M_d, N_d, P_d, Width);
